+++ cmd/gostack/main.go
@@ -51,6 +51,7 @@ func main() {
 	kernel := console.NewKernel()
 	kernel.Register(&console.NewCommand{})
 	kernel.Register(&console.ServeCommand{})
+	kernel.Register(&console.DevCommand{})
 	kernel.Register(&console.LangSearchCommand{})
 	kernel.Register(&console.MigrateCommand{})
 	kernel.Register(&console.RollbackCommand{})